refactor supabase helpers
